refactor : Controller와 Restdocs를 하나로 통합하여 테스트로 변경
